* commit '8ecc895be789040e410e8fca7d9bbc465777a0b7':
  Squashed 'helpers/' changes from e0de48bcc..13cc21a65
  VFS-4262 Updated asio executors to new API
  Squashed 'helpers/' changes from dc4f4562e..e0de48bcc

+++ src/cache/helpersCache.cc
@@ -34,7 +34,8 @@ HelpersCache::HelpersCache(communication::Communicator &communicator,
     : m_communicator{communicator}
     , m_scheduler{scheduler}
     , m_options{options}
-    , m_helpersIoService{options.getStorageHelperThreadCount()}
+    , m_helpersIoService{static_cast<int>(
+          options.getStorageHelperThreadCount())}
     , m_helperFactory
 {
 #if WITH_CEPH
